Update Hashcheck.bat
Updated to be used as a command instead. Instead of double clicking it and entering the full file path with no tab autocompletion or copy/paste, you now use Hashcheck.bat as a command and the filepath becomes the argument.

Usage: C:\Users\Username> Hashcheck.bat C:\Users\Username\whateveryouwant.txt 
Hashcheck.bat can only be invoked if you are operating in the directory in which it is stored. Use DIR and CD commands respectively to navigate the file system from CMD.